removed old firebase setup and replaced it with redux-firestore / react-redux-firebase implementations
