+++ src/config/firebase.ts
@@ -1,33 +1,26 @@
-import { initializeApp } from 'firebase/app';
-import { getFirestore } from 'firebase/firestore';
-import 'firebase/auth';
-import {
-    getAuth,
-    GoogleAuthProvider,
-    FacebookAuthProvider,
-    TwitterAuthProvider,
-    UserCredential,
-    signInWithPopup
-} from 'firebase/auth';
-import firebaseConfiguration from './firebaseConfig';
+import { applyMiddleware, createStore } from 'redux';
+import { rootReducer } from '../reducers';
+import { composeWithDevTools } from 'redux-devtools-extension';
+import thunk from 'redux-thunk';
+import firebase from 'firebase/compat';
+import firebaseConfig from './firebaseConfig';
+import { createFirestoreInstance } from 'redux-firestore';
+import { getAuth } from 'firebase/auth';
 
-const Firebase = initializeApp(firebaseConfiguration);
-
-export const auth = getAuth();
-export const db = getFirestore();
-
-export const Providers = {
-    google: new GoogleAuthProvider(),
-    facebook: new FacebookAuthProvider(),
-    twitter: new TwitterAuthProvider()
+const rrfConfig = {
+    userProfile: 'users',
+    useFirestoreForProfile: true
 };
 
-Providers.google.setCustomParameters({ prompt: 'select_account' });
-Providers.facebook.setCustomParameters({ display: 'popup' });
-Providers.twitter.setCustomParameters({ prompt: 'select_account' });
+const initialState = {};
+export const store = createStore(rootReducer, initialState, composeWithDevTools(applyMiddleware(thunk)));
+// https://stackoverflow.com/questions/68946446/how-do-i-fix-a-firebase-9-0-import-error-attempted-import-error-firebase-app/68967024#68967024
+firebase.initializeApp(firebaseConfig);
 
-export const signInWithGoogle = (): Promise<UserCredential> => signInWithPopup(auth, Providers.google);
-export const signInWithFacebook = (): Promise<UserCredential> => signInWithPopup(auth, Providers.facebook);
-export const signInWithTwitter = (): Promise<UserCredential> => signInWithPopup(auth, Providers.twitter);
-
-export default Firebase;
+export const auth = getAuth();
+export const rrfProps = {
+    firebase,
+    config: rrfConfig,
+    dispatch: store.dispatch,
+    createFirestoreInstance
+};
