+++ src/index.tsx
@@ -1,4 +1,4 @@
-import { Suspense, StrictMode } from 'react';
+import { StrictMode, Suspense } from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
@@ -7,40 +7,15 @@ import { BrowserRouter } from 'react-router-dom';
 import './i18n';
 
 import { Provider } from 'react-redux';
-import { createStore, applyMiddleware } from 'redux';
-import { composeWithDevTools } from 'redux-devtools-extension';
-
-import thunk from 'redux-thunk';
 import { SnackbarProvider } from 'notistack';
 import AuthContextProvider from './contexts/AuthContextProvider';
 import { ThemeProvider } from '@mui/material';
 import theme from './theme/theme';
-// import Firebase from './config/firebase';
-import firebase from 'firebase/compat/app';
 import 'firebase/compat/auth';
 import 'firebase/compat/firestore';
 
-import { createFirestoreInstance } from 'redux-firestore';
 import { ReactReduxFirebaseProvider } from 'react-redux-firebase';
-import { rootReducer } from './reducers';
-import firebaseConfig from './config/firebaseConfig';
-
-const rrfConfig = {
-    userProfile: 'users',
-    useFirestoreForProfile: true
-};
-
-const initialState = {};
-const store = createStore(rootReducer, initialState, composeWithDevTools(applyMiddleware(thunk)));
-// https://stackoverflow.com/questions/68946446/how-do-i-fix-a-firebase-9-0-import-error-attempted-import-error-firebase-app/68967024#68967024
-firebase.initializeApp(firebaseConfig);
-
-const rrfProps = {
-    firebase,
-    config: rrfConfig,
-    dispatch: store.dispatch,
-    createFirestoreInstance
-};
+import { rrfProps, store } from './config/firebase';
 
 ReactDOM.render(
     <Suspense fallback={<></>}>
