+++ src/pages/authentication/RegistrationPage.tsx
@@ -1,5 +1,5 @@
 import { FC, useState } from 'react';
-import { createUserWithEmailAndPassword } from 'firebase/auth';
+// import { createUserWithEmailAndPassword } from 'firebase/auth';
 import PasswordField from '../../components/PasswordField';
 import { useHistory } from 'react-router-dom';
 // import TextField from '../../components/TextField';
@@ -8,7 +8,7 @@ import logger from '../../logging/logger';
 import AuthContainer from '../../components/AuthContainer';
 import { useTranslation } from 'react-i18next';
 import { useSnackbar } from 'notistack';
-import { auth } from '../../config/firebase';
+// import { auth } from '../../config/firebase';
 
 const RegistrationPage: FC = () => {
     const { t } = useTranslation();
@@ -27,23 +27,23 @@ const RegistrationPage: FC = () => {
         }
         // setRegistering(true);
 
-        createUserWithEmailAndPassword(auth, email, password)
-            .then((result) => {
-                // logging.info(result);
-                console.log(result);
-                // setRegistering(false);
-                history.push('/login');
-            })
-            .catch((error) => {
-                console.error(error.code);
-                // 'auth/weak-password'
-                if (error.code === 'auth/email-already-in-use') {
-                    enqueueSnackbar('Email Already In Use', { variant: 'error' });
-                }
-
-                // TODO : error handling...
-                // setRegistering(false);
-            });
+        // createUserWithEmailAndPassword(auth, email, password)
+        //     .then((result) => {
+        //         // logging.info(result);
+        //         console.log(result);
+        //         // setRegistering(false);
+        //         history.push('/login');
+        //     })
+        //     .catch((error) => {
+        //         console.error(error.code);
+        //         // 'auth/weak-password'
+        //         if (error.code === 'auth/email-already-in-use') {
+        //             enqueueSnackbar('Email Already In Use', { variant: 'error' });
+        //         }
+        //
+        //         // TODO : error handling...
+        //         // setRegistering(false);
+        //     });
     };
 
     return (
